Chore - move telemetry to be an app level service
