+++ plugins/amazonq/codetransform/jetbrains-community/src/software/aws/toolkits/jetbrains/services/codemodernizer/CodeModernizerManager.kt
@@ -40,7 +40,7 @@ import software.aws.toolkits.jetbrains.services.codemodernizer.model.CodeModerni
 import software.aws.toolkits.jetbrains.services.codemodernizer.model.CodeModernizerSessionContext
 import software.aws.toolkits.jetbrains.services.codemodernizer.model.CodeModernizerStartJobResult
 import software.aws.toolkits.jetbrains.services.codemodernizer.model.CodeTransformHilDownloadArtifact
-import software.aws.toolkits.jetbrains.services.codemodernizer.model.CodeTransformTelemetryMetadataSingleton
+import software.aws.toolkits.jetbrains.services.codemodernizer.model.CodeTransformTelemetryService
 import software.aws.toolkits.jetbrains.services.codemodernizer.model.CustomerSelection
 import software.aws.toolkits.jetbrains.services.codemodernizer.model.Dependency
 import software.aws.toolkits.jetbrains.services.codemodernizer.model.InvalidTelemetryReason
@@ -121,7 +121,7 @@ class CodeModernizerManager(private val project: Project) : PersistentStateCompo
     )
     init {
         CodeModernizerSessionState.getInstance(project).setDefaults()
-        CodeTransformTelemetryMetadataSingleton.getInstance().resetDefaults()
+        CodeTransformTelemetryService.getInstance().resetDefaults()
     }
 
     fun validate(project: Project): ValidationResult {
