+++ plugins/amazonq/codetransform/jetbrains-community/src/software/aws/toolkits/jetbrains/services/codemodernizer/controller/CodeTransformChatController.kt
@@ -58,7 +58,7 @@ import software.aws.toolkits.jetbrains.services.codemodernizer.messages.CodeTran
 import software.aws.toolkits.jetbrains.services.codemodernizer.messages.IncomingCodeTransformMessage
 import software.aws.toolkits.jetbrains.services.codemodernizer.model.CodeModernizerJobCompletedResult
 import software.aws.toolkits.jetbrains.services.codemodernizer.model.CodeTransformHilDownloadArtifact
-import software.aws.toolkits.jetbrains.services.codemodernizer.model.CodeTransformTelemetryMetadataSingleton
+import software.aws.toolkits.jetbrains.services.codemodernizer.model.CodeTransformTelemetryService
 import software.aws.toolkits.jetbrains.services.codemodernizer.model.CustomerSelection
 import software.aws.toolkits.jetbrains.services.codemodernizer.model.DownloadFailureReason
 import software.aws.toolkits.jetbrains.services.codemodernizer.model.JobId
@@ -417,10 +417,10 @@ class CodeTransformChatController(
 
             codeModernizerManager.resumePollingFromHil()
         } catch (e: Exception) {
-            CodeTransformTelemetryMetadataSingleton.setCancelledFromChat(false)
+            CodeTransformTelemetryService.setCancelledFromChat(false)
             telemetry.logHil(
                 CodeModernizerSessionState.getInstance(context.project).currentJobId?.id.orEmpty(),
-                CodeTransformTelemetryMetadataSingleton.getInstance(),
+                CodeTransformTelemetryService.getInstance(),
                 success = false,
                 reason = "Runtime Error when trying to resume transformation from HIL",
             )
@@ -509,10 +509,10 @@ class CodeTransformChatController(
 
         try {
             codeModernizerManager.tryResumeWithAlternativeVersion(selectedVersion)
-            CodeTransformTelemetryMetadataSingleton.setDependencyVersionSelected(selectedVersion)
+            CodeTransformTelemetryService.setDependencyVersionSelected(selectedVersion)
             telemetry.logHil(
                 CodeModernizerSessionState.getInstance(context.project).currentJobId?.id as String,
-                CodeTransformTelemetryMetadataSingleton.getInstance(),
+                CodeTransformTelemetryService.getInstance(),
                 success = true,
                 reason = "User selected version"
             )
@@ -539,10 +539,10 @@ class CodeTransformChatController(
 
         try {
             codeModernizerManager.rejectHil()
-            CodeTransformTelemetryMetadataSingleton.setCancelledFromChat(true)
+            CodeTransformTelemetryService.setCancelledFromChat(true)
             telemetry.logHil(
                 CodeModernizerSessionState.getInstance(context.project).currentJobId?.id.orEmpty(),
-                CodeTransformTelemetryMetadataSingleton.getInstance(),
+                CodeTransformTelemetryService.getInstance(),
                 success = false,
                 reason = "User cancelled"
             )
@@ -552,10 +552,10 @@ class CodeTransformChatController(
             }
             codeModernizerManager.resumePollingFromHil()
         } catch (e: Exception) {
-            CodeTransformTelemetryMetadataSingleton.setCancelledFromChat(false)
+            CodeTransformTelemetryService.setCancelledFromChat(false)
             telemetry.logHil(
                 CodeModernizerSessionState.getInstance(context.project).currentJobId?.id.orEmpty(),
-                CodeTransformTelemetryMetadataSingleton.getInstance(),
+                CodeTransformTelemetryService.getInstance(),
                 success = false,
                 reason = "Runtime Error when trying to resume transformation from HIL",
             )
